# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Java Speaker Club API
  description: API для управления докладами, лекциями, ролями и медиафайлами спикеров
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/auth/sign-on:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Пользователь зарегистрирован и получил токены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          description: Ошибка регистрации (например, username уже занят)
  /api/auth/sign-in:
    post:
      tags: [Auth]
      summary: Аутентификация пользователя
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          description: Неверные учётные данные
...
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Выход из системы
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refresh-токен отозван
        '401':
          description: Недействительный refresh-токен
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Обновить пару access/refresh токенов
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          description: Refresh-токен недействителен или отозван

  # ── Profile ─────────────────────────────────────────────────────────
...
  /api/roles:
    get:
      tags: [Roles]
      summary: Список всех ролей
      operationId: findAllRoles
      responses:
        '200':
          description: Список ролей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'

    post:
      tags: [Roles]
      summary: Создать роль
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: Роль создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Спикер не найден или роль уже назначена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/media/upload:
    post:
      tags: [Media]
      summary: Загрузить медиафайл без привязки к лекции
      operationId: uploadStandaloneMedia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Загружаемый файл
      responses:
        '201':
          description: Файл загружен, создана запись с идентификатором и ключом MinIO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMediaUploadResponse'
        '400':
          description: Ошибка загрузки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/media/upload/{lectureId}:
    post:
      tags: [Media]
      summary: Загрузить медиафайл для лекции
      operationId: uploadMedia
      parameters:
        - $ref: '#/components/parameters/MediaLectureId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Загружаемый файл
      responses:
        '201':
          description: Медиафайл загружен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LectureResponse'
        '400':
          description: Ошибка загрузки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/media/download/{lectureId}:
    get:
      tags: [Media]
      summary: Скачать медиафайл лекции
      operationId: downloadMedia
      parameters:
        - $ref: '#/components/parameters/MediaLectureId'
      responses:
        '200':
          description: Файл для скачивания
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="presentation.pdf"'
        '404':
          description: Лекция не найдена или медиафайл отсутствует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/media/url/{lectureId}:
    get:
      tags: [Media]
      summary: Получить presigned URL медиафайла
      operationId: getPresignedUrl
      parameters:
        - $ref: '#/components/parameters/MediaLectureId'
      responses:
        '200':
          description: Presigned URL
          content:
            text/plain:
              schema:
                type: string
                example: https://minio:9000/lectures-media/...?X-Amz-Signature=...
        '404':
          description: Лекция не найдена или медиафайл отсутствует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/media/{lectureId}:
    delete:
      tags: [Media]
      summary: Удалить медиафайл лекции
      operationId: deleteMedia
      parameters:
        - $ref: '#/components/parameters/MediaLectureId'
      responses:
        '204':
          description: Медиафайл удалён
        '404':
          description: Лекция не найдена или медиафайл отсутствует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/reports/filters:
    get:
      tags: [Reports]
      summary: Получить справочники и опции для фильтров дашбордов
      operationId: getReportFilters
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Справочники для speaker и DevRel dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportFiltersResponse'
        '401':
          description: Пользователь не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/reports/speaker-dashboard/generate:
    post:
      tags: [Reports]
      summary: Сгенерировать JSON-отчёт для дашборда спикера
      operationId: generateSpeakerDashboardReport
      description: Возвращает данные для личного кабинета спикера, включая графики, достижения и рекомендации CFP.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakerDashboardReportRequest'
      responses:
        '200':
          description: Отчёт для speaker dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerDashboardReport'
        '401':
          description: Пользователь не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Спикер не найден или для него ещё нет данных отчёта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/reports/devrel-dashboard/generate:
    post:
      tags: [Reports]
      summary: Сгенерировать JSON-отчёт для дашборда DevRel
      operationId: generateDevRelDashboardReport
      description: Возвращает агрегаты, рейтинги, blind spots и поток карточек для парсинга мероприятий.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevRelDashboardReportRequest'
      responses:
        '200':
          description: Отчёт для DevRel dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevRelDashboardReport'
        '401':
          description: Пользователь не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Недостаточно прав для DevRel-аналитики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  parameters:
...
    MediaLectureId:
      name: lectureId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: johndoe
        password:
          type: string
          format: password
          example: secret123
    RoleName:
      type: string
      enum: [ADMIN, DEVREL, SPEAKER, USER]
    AuthResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: johndoe
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleName'
        createdDate:
          type: string
          format: date-time
        lastModifiedDate:
          type: string
          format: date-time
    AuthTokensResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access-токен
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: Время жизни access в секундах
        refreshToken:
          type: string
        refreshTokenExpiresIn:
          type: integer
          format: int64
          description: Время жизни refresh в секундах
        user:
          $ref: '#/components/schemas/AuthResponse'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    LectureResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        talkId:
          type: string
          format: uuid
        speakerId:
          type: integer
          format: int64
        hasMedia:
          type: boolean
          description: Есть ли прикрепленный медиафайл
        mediaFilename:
          type: string
          nullable: true
          description: Оригинальное имя файла
          example: presentation.pdf
        mediaMime:
          type: string
          nullable: true
          description: MIME тип медиафайла (если доступен)
          example: application/pdf
        createdDate:
          type: string
          format: date-time
    UserMediaUploadResponse:
      type: object
      required: [id, mediaMinioKey]
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор записи о загрузке
        mediaMinioKey:
          type: string
          description: Уникальный ключ объекта в MinIO

    # ── Roles ───────────────────────────────────────────────────────
    RoleCreateRequest:
      type: object
      required: [name, speakerId]
      properties:
        name:
          type: string
          example: SPEAKER
        speakerId:
          type: integer
          format: int64
          example: 1
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: SPEAKER
        speakerId:
          type: integer
          format: int64
        createdDate:
          type: string
          format: date-time

    # ── Speakers ────────────────────────────────────────────────────
...
    SelectOption:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
...
    RoleFilterGroup:
      type: object
      required: [label, key, children]
      properties:
        label:
          type: string
        key:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/RoleFilterOption'
    ReportFiltersResponse:
      type: object
      required:
        - periods
        - regions
        - companies
        - roles
        - influences
        - tagWindows
        - parsingSources
        - parsingStatuses
      properties:
        periods:
          type: array
          items:
            $ref: '#/components/schemas/SelectOption'
        regions:
          type: array
          items:
            type: string
        companies:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleFilterGroup'
        influences:
          type: array
          items:
            $ref: '#/components/schemas/SelectOption'
        tagWindows:
          type: array
          items:
            $ref: '#/components/schemas/SelectOption'
        parsingSources:
          type: array
          items:
            $ref: '#/components/schemas/SelectOption'
        parsingStatuses:
          type: array
          items:
            $ref: '#/components/schemas/SelectOption'
    DashboardPeriod:
      type: string
      enum: ['30d', quarter, year, all]
    DashboardInfluence:
      type: string
      enum: [local, regional, federal, international]
    ParsingStatus:
      type: string
      enum: [new, duplicate, confirmed, rejected]
    TagWindow:
      type: string
      enum: [week, month]
    TrendSeries:
      type: object
      required: [labels, talks, nps]
      properties:
        labels:
          type: array
          items:
            type: string
        talks:
          type: array
          items:
            type: integer
            minimum: 0
        nps:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 100
...
    RecommendationCard:
      type: object
      required: [id, title, city, level, topic, deadline, match, hostCompany, url]
      properties:
        id:
          type: string
        title:
          type: string
        city:
          type: string
        level:
          $ref: '#/components/schemas/DashboardInfluence'
        topic:
          type: string
        deadline:
          type: string
          format: date
        match:
          type: integer
          minimum: 0
          maximum: 100
        hostCompany:
          type: string
        url:
          type: string
          format: uri
...
    TagCloudItem:
      type: object
      required: [label, weight]
      properties:
        label:
          type: string
        weight:
          type: integer
          minimum: 1
    DashboardSpeaker:
      type: object
      required:
        - id
        - name
        - username
        - role
        - roleKey
        - roleGroup
        - roleGroupKey
        - region
        - company
        - influence
        - nps
        - previousNps
        - talksCount
        - lastTalkDate
        - topConferenceCandidate
        - tags
        - profileCompletion
        - lastEvaluationDate
        - audience
        - talkMix
        - series
        - heatmap
        - achievements
        - recommendations
        - avatarColor
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        role:
          type: string
        roleKey:
          type: string
        roleGroup:
          type: string
        roleGroupKey:
          type: string
        region:
          type: string
        company:
          type: string
        influence:
          $ref: '#/components/schemas/DashboardInfluence'
        nps:
          type: integer
          minimum: 0
          maximum: 100
        previousNps:
          type: integer
          minimum: 0
          maximum: 100
        talksCount:
          type: integer
          minimum: 0
        lastTalkDate:
          type: string
          format: date-time
        topConferenceCandidate:
          type: boolean
        tags:
          type: array
          items:
            type: string
        profileCompletion:
          type: integer
          minimum: 0
          maximum: 100
        lastEvaluationDate:
          type: [string, 'null']
          format: date-time
        audience:
          type: integer
          minimum: 0
        talkMix:
          $ref: '#/components/schemas/InfluenceBreakdown'
        series:
          $ref: '#/components/schemas/DashboardSeriesByPeriod'
        heatmap:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 4
        achievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/RecommendationCard'
        avatarColor:
          type: string
          example: '#0ea5e9'
...
    SpeakerDashboardReportRequest:
      type: object
      additionalProperties: false
      properties:
        period:
          $ref: '#/components/schemas/DashboardPeriod'
        targetRegions:
          type: array
          items:
            type: string
        targetRoles:
          type: array
          items:
            type: string
        targetCompanies:
          type: array
          items:
            type: string
        targetInfluence:
          oneOf:
            - $ref: '#/components/schemas/DashboardInfluence'
            - type: 'null'
        npsMin:
          type: integer
          minimum: 0
          maximum: 100
          description: Нижняя граница фильтра рекомендаций по score match
        npsMax:
          type: integer
          minimum: 0
          maximum: 100
          description: Верхняя граница фильтра рекомендаций по score match
    SpeakerDashboardSummary:
      type: object
      required:
        - talksCount
        - lastTalkDate
        - nps
        - previousNps
        - trendDirection
        - trendDelta
        - audience
        - activityGrowth
        - activeTopicsCount
        - relevantRecommendationsCount
      properties:
        talksCount:
          type: integer
          minimum: 0
        lastTalkDate:
          type: string
          format: date-time
        nps:
          type: integer
          minimum: 0
          maximum: 100
        previousNps:
          type: integer
          minimum: 0
          maximum: 100
        trendDirection:
          type: string
          enum: [up, down, stable]
        trendDelta:
          type: integer
        audience:
          type: integer
          minimum: 0
        activityGrowth:
          type: integer
        activeTopicsCount:
          type: integer
          minimum: 0
        relevantRecommendationsCount:
          type: integer
          minimum: 0
    SpeakerDashboardReport:
      type: object
      required:
        - generatedAt
        - filtersApplied
        - speaker
        - currentSeries
        - summary
        - tagCloud
        - filteredRecommendations
      properties:
        generatedAt:
          type: string
          format: date-time
        filtersApplied:
          $ref: '#/components/schemas/SpeakerDashboardReportRequest'
        speaker:
          $ref: '#/components/schemas/DashboardSpeaker'
        currentSeries:
          $ref: '#/components/schemas/TrendSeries'
        summary:
          $ref: '#/components/schemas/SpeakerDashboardSummary'
        tagCloud:
          type: array
          items:
            $ref: '#/components/schemas/TagCloudItem'
        filteredRecommendations:
          type: array
          items:
            $ref: '#/components/schemas/RecommendationCard'
    DevRelDashboardReportRequest:
      type: object
      additionalProperties: false
      properties:
        period:
          $ref: '#/components/schemas/DashboardPeriod'
        regions:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        companies:
          type: array
          items:
            type: string
        influence:
          oneOf:
            - $ref: '#/components/schemas/DashboardInfluence'
            - type: 'null'
        npsMin:
          type: integer
          minimum: 0
          maximum: 100
        npsMax:
          type: integer
          minimum: 0
          maximum: 100
        speakerSearch:
          type: string
        tagWindow:
          $ref: '#/components/schemas/TagWindow'
        parsingSource:
          type: [string, 'null']
        parsingStatus:
          oneOf:
            - $ref: '#/components/schemas/ParsingStatus'
            - type: 'null'
    DevRelDashboardSummary:
      type: object
      required: [activeSpeakers, avgNps, risingSpeakers, confirmedRate, topConferenceCandidates]
      properties:
        activeSpeakers:
          type: integer
          minimum: 0
        avgNps:
          type: integer
          minimum: 0
          maximum: 100
        risingSpeakers:
          type: integer
          minimum: 0
        confirmedRate:
          type: integer
          minimum: 0
          maximum: 100
        topConferenceCandidates:
          type: integer
          minimum: 0
    DevRelRatingSection:
      type: object
      required: [speakers, topConferenceCandidates, averageTalks, internationalSpeakers]
      properties:
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/DashboardSpeaker'
        topConferenceCandidates:
          type: array
          items:
            $ref: '#/components/schemas/DashboardSpeaker'
        averageTalks:
          type: integer
          minimum: 0
        internationalSpeakers:
          type: integer
          minimum: 0
    DevRelAnalyticsSection:
      type: object
      required: [aggregateSeries, influenceMix, tagCloud, dominantTag]
      properties:
        aggregateSeries:
          $ref: '#/components/schemas/TrendSeries'
        influenceMix:
          $ref: '#/components/schemas/InfluenceBreakdown'
        tagCloud:
          type: array
          items:
            $ref: '#/components/schemas/TagCloudItem'
        dominantTag:
          type: string
    DevRelBlindSpotsSection:
      type: object
      required: [unevaluatedShare, inactiveSpeakersCount, incompleteProfilesCount, rows]
      properties:
        unevaluatedShare:
          type: integer
          minimum: 0
          maximum: 100
        inactiveSpeakersCount:
          type: integer
          minimum: 0
        incompleteProfilesCount:
          type: integer
          minimum: 0
        rows:
          type: array
          items:
            $ref: '#/components/schemas/BlindSpotRow'
...
    DevRelDashboardReport:
      type: object
      required: [generatedAt, filtersApplied, summary, rating, analytics, blindSpots, parsing]
      properties:
        generatedAt:
          type: string
          format: date-time
        filtersApplied:
          $ref: '#/components/schemas/DevRelDashboardReportRequest'
        summary:
          $ref: '#/components/schemas/DevRelDashboardSummary'
        rating:
          $ref: '#/components/schemas/DevRelRatingSection'
        analytics:
          $ref: '#/components/schemas/DevRelAnalyticsSection'
        blindSpots:
          $ref: '#/components/schemas/DevRelBlindSpotsSection'
        parsing:
          $ref: '#/components/schemas/DevRelParsingSection'
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: "Talk not found: 550e8400-e29b-41d4-a716-446655440000"
        instance:
          type: string
          format: uri